requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
flex==6.14.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonpointer==3.0.0
openai==1.35.14
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rfc3987==1.3.8
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
strict-rfc3339==0.7
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
validate-email==1.3
